% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{compound}
\alias{compound}
\title{Generate compound symmetric matrices}
\usage{
compound(p = 8, n = NULL)
}
\arguments{
\item{p}{desired dimension}

\item{n}{option to generate n observations from covariance matrix S}
}
\value{
Omega, S
}
\description{
Generate a p-dimensional compound symmetric matrix.
}
\keyword{internal}
