% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{dense}
\alias{dense}
\title{Generate dense matrices}
\usage{
dense(p = 8, base = 0.9, n = NULL)
}
\arguments{
\item{p}{desired dimension}

\item{base}{base multiplier}

\item{n}{option to generate n observations from covariance matrix S}
}
\value{
Omega, S
}
\description{
Generate p-dimensional matrices so that its inverse is dense.
}
\keyword{internal}
