% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{denseQR}
\alias{denseQR}
\title{Generate dense matrices (via spectral decomposition)}
\usage{
denseQR(p = 8, num = 5, n = NULL)
}
\arguments{
\item{p}{desired dimension}

\item{num}{number of 'large' eigen values. Note num must be less than p}

\item{n}{option to generate n observations from covariance matrix S}
}
\value{
Omega, S
}
\description{
Generate p-dimensional matrices so that its inverse is dense. The matrix will be generated so its first 'num' eigen values are 1000 and the remaining are 1. The orthogonal basis is generated via QR decomposition of
}
\keyword{internal}
