% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{plot.shrink}
\alias{plot.shrink}
\title{Plot shrink object}
\usage{
\method{plot}{shrink}(x, type = c("line", "heatmap"), footnote = TRUE, ...)
}
\arguments{
\item{x}{class object shrink.}

\item{type}{produce either 'heatmap' or 'line' graph}

\item{footnote}{option to print footnote of optimal values. Defaults to TRUE.}

\item{...}{additional arguments.}
}
\description{
Produces a plot for the cross validation errors, if available.
}
\examples{
# NEED TO COMPLETE
}
