% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{print.shrink}
\alias{print.shrink}
\title{Print shrink object}
\usage{
\method{print}{shrink}(x, ...)
}
\arguments{
\item{x}{class object shrink}

\item{...}{additional arguments.}
}
\description{
Prints shrink object and suppresses output if needed.
}
\keyword{internal}
