% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{tridiag}
\alias{tridiag}
\title{Generate tri-diagonal matrices}
\usage{
tridiag(p = 8, base = 0.7, n = NULL)
}
\arguments{
\item{p}{desired dimension}

\item{base}{base multiplier}

\item{n}{option to generate n observations from covariance matrix S}
}
\value{
Omega, S
}
\description{
Generate p-dimensional matrices so that its inverse is tri-diagonal.
}
\keyword{internal}
